Fix infinite loop in client_echo example
Previously username was an object that contained the message that the server sent the client. When JavaScript checked to see if username (an object) === client.username (a string) it always returned false, thus entering into an infinite loop as the chat client responded to itself indefinitely. This patch addresses by parsing out the username from the object.